Add script to 'cat' with syntax highlighting.
It's both a valid bash script and vim script.  When executed as a bash
script, it sets up a file descriptor on /proc/self/fd/9 and executes vim
to source this file as well.  Vim calculates the colors, writes the file
complete with ANSI color escapes out to /proc/self/fd/9, and exits; and
finally bash exec's a cat to read from fd 9.