Several small changes for :source / GenericHandler
Before calling any handlers, set g:netlib_operation to either 'read',
'write', 'append', or 'source' so that plugins that aren't suitable for
one of those operations can bail.  Use it for http dumpers, so that we
only try :source'ing the result of a source-dumper and not a page-dumper

Change GenericHandler to accept a vimscript expression to check if
everything needed is installed, to make using the above variable possible.