More work towards sane tempfile names for writes
One function sets the tempfile name based upon g:netlib_operation.
Largely the same logic as before is followed, except that the name for
writes will never match the actual buffer name - it will be the last
segment of the 'path' portion of the URI, or 'write_temp' if we cannot
split the URI.

Handlers are now responsible for ensuring that the temp file name
matches the final directory component and stripping that component
themselves where necessary.  This needs to be done on a per-handler
basis, since it was wrong for some handlers (eg file://).  This is
assisted by the addition of netlib#utility#get_directory_path().