2 plugins for fixed source form Fortran editing
fortran-indenting.vim provides normal and visual mode overrides for >>
and > to prevent them from modifying the first 6 columns of the file
(unless a comment marker is set in the first column, in which case only
the 1st column will be prevented from shifting).

functionwise-jumps.vim provides overrides for [[ [] ][ ]] to jump to the
start or end of any subroutine, entry, function, or program unit.